requirements.txt
alabaster==0.7.10
astroid==1.6.2
attrs==17.4.0
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
docutils==0.14
enum34==1.1.6
Faker==0.8.12
idna==2.6
imagesize==1.0.0
jsonpickle==0.9.6
isort==4.3.4
Jinja2==2.10
jsonpickle==0.9.6
junit2html==20
lazy-object-proxy==1.3.1
lxml==4.2.0
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
mysqlclient==1.3.12
namedlist==1.7
packaging==17.1
pbr==3.1.1
pluggy==0.6.0
py==1.5.2
Pygments==2.2.0
pylint==1.8.3
pyparsing==2.2.0
pytest==3.4.2
pytest-allure-adaptor==1.7.10
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
selenium==3.11.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinx-rtd-theme==0.3.0
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.2.5
text-unidecode==1.2
urllib3==1.22
wrapt==1.10.11
